+++ pkg/providers/appsec/testdata/ase-botman/modules/security/advanced-settings.tf
@@ -1,0 +1,48 @@
+resource "akamai_botman_bot_analytics_cookie" "bot_analytics_cookie" {
+  config_id = akamai_appsec_configuration.config.config_id
+  bot_analytics_cookie = jsonencode(
+    {
+      "arrayKey" : [
+        "arrayValueA1",
+        "arrayValueA2"
+      ],
+      "objectKey" : {
+        "innerKey" : "innerValueA"
+      },
+      "primitiveKey" : "primitiveValueA"
+    }
+  )
+}
+
+resource "akamai_botman_client_side_security" "client_side_security" {
+  config_id = akamai_appsec_configuration.config.config_id
+  client_side_security = jsonencode(
+    {
+      "arrayKey" : [
+        "arrayValueA1",
+        "arrayValueA2"
+      ],
+      "objectKey" : {
+        "innerKey" : "innerValueA"
+      },
+      "primitiveKey" : "primitiveValueA"
+    }
+  )
+}
+
+resource "akamai_botman_transactional_endpoint_protection" "transactional_endpoint_protection" {
+  config_id = akamai_appsec_configuration.config.config_id
+  transactional_endpoint_protection = jsonencode(
+    {
+      "arrayKey" : [
+        "arrayValueA1",
+        "arrayValueA2"
+      ],
+      "objectKey" : {
+        "innerKey" : "innerValueA"
+      },
+      "primitiveKey" : "primitiveValueA"
+    }
+  )
+}
+