+++ pkg/providers/appsec/testdata/ase/appsec-import.sh
@@ -201,11 +201,8 @@ terraform import module.security.akamai_appsec_ip_geo.default_policy 79947:ASE1_
 terraform import module.security.akamai_appsec_penalty_box.default_policy 79947:ASE1_156138
 terraform import module.security.akamai_appsec_slow_post.default_policy 79947:ASE1_156138
 terraform import module.security.akamai_appsec_advanced_settings_request_body.default_policy 79947:ASE1_156138
-
 terraform import module.security.akamai_appsec_rate_policy.page_view_requests 79947:177906
-
 terraform import module.security.akamai_appsec_rate_policy.origin_error 79947:177907
-
 terraform import module.security.akamai_appsec_rate_policy.post_page_requests 79947:177908
 terraform import module.security.akamai_appsec_reputation_profile.dos_attackers_high_threat 79947:3017089
 terraform import module.security.akamai_appsec_reputation_profile.dos_attackers_low_threat 79947:3017090